requirements.txt
discord.py==2.1.0
python-dotenv==0.21.0
openai==0.25.0
pnglatex==1.1